Add support for dynamic tags
There was a need to push some other tags that dependend on the request
object. CUrrently, the only way to pass `tags` was during the initial
set-up.

The proposed change makes `tags` options be able to take a `function`.
The passed function takes the `request` object when called, and returns
any additional tags that may be generated dynamically from properties
found in the request object.